refactor: extra types and utils to root of the package
