--- packages/nextjs-mf/src/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './internal';
-export * from './utils';
